refactor: splits cuda oscilation calculation from vbo update kernels
There is now a kernel to update vbo resources, separated from the one that oscilates the data. This allows us to work with padded data in kernels, so we don't have to worry about the threads that are outside the original bounds. The kernel that updates the vbo then cares about the padding.