fix!: index calculation
Half of the code that computed indices in the grid was using something like id=x*maxX + y instead of id=x*maxY + y (since maxX/maxY were called m/n or width/height depending on the code section). This was fixed, so now all the indices are computed the same.